refactor, fix: now the script runs (previously it didn't, heavily refactored across multiple files
