+++ src/__init__.py
@@ -1,0 +1,3 @@
+from .board import *
+from .cells import *
+from .player import *
